pyproject.toml
[project]
name = "colony-manager"
description = "Database schema and optional GUI for colony management"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "license.txt"}
authors = [
    {name = "Brad Buran", email="buran@ohsu.edu"},
]
maintainers = [
    {name = "Brad Buran", email="buran@ohsu.edu"},
]
dependencies = [
    "sqlalchemy",
    "sqlalchemy_continuum",
	"psycopg2-binary",
]
dynamic = ["version"]

[project.optional-dependencies] 
gui = [
	# Not sure if all of these are needed. Maybe legacy cruft.
    "werkzeug",
	"Flask>=2.0",
	"Flask-SQLAlchemy>=2.5",
	"Flask-WTF>=1.0",
	"flask-login",
	"python-dotenv>=0.19",
	"WTForms[email]>=3.0",
	"wtforms-sqlalchemy",
	"Pillow>=10.0",
]

[build-system]
requires = ["setuptools>=61.2", "wheel", "setuptools_scm[toml]>=3.4.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "src/colony_manager/version.py"

[tool.setuptools.packages.find]
where = ["src"]
